refactor: extracted form_group
Removed form_group from form_builder into a separate concern called 
form_group.
Extracted label to it's own file in a components directory.
