+++ lib/tailwind_form/components.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module TailwindForm
+  module Components
+    extend ActiveSupport::Autoload
+
+    autoload :Labels
+
+    include Labels
+  end
+end
