+++ lib/tailwind_form/form_builder.rb
@@ -2,23 +2,14 @@
 
 module TailwindForm
   class FormBuilder < ActionView::Helpers::FormBuilder
+    include TailwindForm::FormGroup
+    include TailwindForm::Components
+
     include TailwindForm::Inputs::Base
     include TailwindForm::Inputs::EmailField
     include TailwindForm::Inputs::TextArea
     include TailwindForm::Inputs::TextField
 
     delegate :concat, :tag, to: :@template
-
-    def tailwind_label_field(name)
-      field_options = { class: 'form__label' }
-      label(name, field_options)
-    end
-
-    def form_group(name, field_type = :text_field, classes = 'form__group')
-      function_name = "#{field_type}_with_tailwind"
-      tag.div(class: classes) do
-        tailwind_label_field(name) + send(function_name, name, { class: 'form__input' })
-      end
-    end
   end
 end
